Multiscreen Apps and Platform API - completed section
converted from a modal process into a Tab-based navigation.  Also added a "location" button that will render the current lat/long of the device.  Note that this will only work in the iOS emulator and not the Android one.  Just launch the app on a real Android device and you will see the lat/long